% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeting.R
\name{greeting}
\alias{greeting}
\title{Forms a greeting}
\usage{
greeting(my_name, major, graduating_year)
}
\arguments{
\item{my_name}{a character containing the input name}

\item{major}{a character containing the input major}

\item{graduating_year}{a number containing year graduated}
}
\value{

}
\description{
Contains one function that forms a greeting for the user
}
